+++ src/index.js
@@ -6,6 +6,7 @@ export {
   compileJSONPointer,
   compileJSONPointerProp,
   parseJSONPointer,
+  parseJSONPointerWithArrayIndices,
   startsWithJSONPointer,
   appendToJSONPointer
 } from './jsonPointer.js'