+++ src/jsonPointer.js
@@ -1,7 +1,9 @@
+import { getIn } from './immutabilityHelpers.js'
+
 /**
  * Parse a JSON Pointer
  * @param {JSONPointer} pointer
- * @return {JSONPath}
+ * @return {string[]}
  */
 export function parseJSONPointer (pointer) {
   const path = pointer.split('/')
@@ -9,6 +11,39 @@ export function parseJSONPointer (pointer) {
 
   return path.map(p => p.replace(/~1/g, '/').replace(/~0/g, '~'))
 }
+
+/**
+ * Parse a JSONPointer, and turn array indices into numeric values.
+ * For example, '/array/2/name' returns ['array', 2, 'name'] when array turns
+ * out to be an actual Array
+ * @param {JSONData} json
+ * @param {JSONPointer} pointer
+ * @return {JSONPath}
+ */
+export function parseJSONPointerWithArrayIndices (json, pointer) {
+  const path = parseJSONPointer(pointer)
+
+  // parse Array indexes into a number
+  for (let i = 0; i < path.length; i++) {
+    const section = path[i]
+
+    if (ARRAY_INDEX_REGEX.exec(section)) {
+      // this path part contains a number.
+      // See if the document actually contains an array
+      const parentPath = path.slice(0, i)
+      const parent = getIn(json, parentPath)
+
+      if (Array.isArray(parent)) {
+        path[i] = parseInt(section, 10)
+      }
+    }
+  }
+
+  return path
+}
+
+// test whether a string only contains one or digits, like "1" or "204"
+const ARRAY_INDEX_REGEX = /^\d+$/
 
 /**
  * Compile a JSON Pointer