+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # Changelog
 
-
-## not yet published, version 2.0.0
+## 2022-06-27, version 2.0.0
 
 - Improved TypeScript definitions (there are some breaking changes).
 - Implemented new util functions: `compileJSONPointerProp`,